Создание слайдера с примененм эффектов вынесено в отдельный файл. Перенесены констаны их form.js в const.js
