+++ js/const.js
@@ -16,8 +16,16 @@ const minCountILikes = 30;
 const maxCountILikes = 200;
 const minCommentsCount = 0;
 const maxCommentsCount = 30;
-
-
 const POSTS_ARRAY_LENGTH = 25; // Длина массива с объектами (по ДЗ - это 25);
 
-export {DESCRIPTION_PHOTOS, TEXT_COMMENTS, USERS_NAME, POSTS_ARRAY_LENGTH, minCountIdPost, maxCountIdPost, minCountIdPhoto, maxCountIdPhoto, minCountIdComment, maxCountIdComment, minCountAvatar, maxCountAvatar, minCountILikes, maxCountILikes, minCommentsCount, maxCommentsCount};
+
+const MAX_HASHTAG_SYMBOLS = 20;
+const MAX_COMMENT_SYMBOLS = 140;
+const MAX_QUANTITY_HASHTAGS = 5;
+const DEFAULT_SCALE_PHOTO = 100;
+const STEP_SCALE_PHOTO = 25;
+const MIN_SCALE_VALUE_PHOTO = 25;
+const MAX_SCALE_VALUE_PHOTO = 100;
+const regexp = /^#[a-zа-яё0-9]+$/i;
+
+export {DESCRIPTION_PHOTOS, TEXT_COMMENTS, USERS_NAME, POSTS_ARRAY_LENGTH, minCountIdPost, maxCountIdPost, minCountIdPhoto, maxCountIdPhoto, minCountIdComment, maxCountIdComment, minCountAvatar, maxCountAvatar, minCountILikes, maxCountILikes, minCommentsCount, maxCommentsCount, MAX_HASHTAG_SYMBOLS, MAX_COMMENT_SYMBOLS, MAX_QUANTITY_HASHTAGS, DEFAULT_SCALE_PHOTO, STEP_SCALE_PHOTO, MIN_SCALE_VALUE_PHOTO,MAX_SCALE_VALUE_PHOTO, regexp};
